api/requirements.txt
accelerate
flask==3.1.1
google-genai
opensearch-py==3.0.0
python-dotenv==1.2.1
sentence-transformers==5.1.2
transformers
